refactor(client): split TheaterPage into focused sub-components
TheaterPage was 234 LOC / 18 functions hosting the data-fetching
wrapper, the date-selection state, the breadcrumb/header/empty-state
JSX, the card grid, and the per-movie meta/ratings/poster markup.

Splits into `components/theater/`:

- TheaterBreadcrumb, TheaterHeader: tiny presentational pieces.
- ShowtimesList (with internal EmptyShowtimes): renders the movie grid
  or the empty-state placeholder.
- MovieShowtimeCard (with internal MovieMetaLine + MovieRatings): the
  poster + title + meta + ratings + showtimes card.
- TheaterContent: the date-selection hook (`useDateTimeFilter`),
  derived memoisation (effectiveSelectedDate, selectedShowtimes,
  movieGroups), and the composition of the above pieces.

TheaterPage becomes a 42-LOC wrapper that loads theaters and the
schedule, computes the not-found / error / loading branches, and
delegates to TheaterContent. The data flow is unchanged: URL params →
`useDateTimeFilter` → memoised filtering → render. All 8 TheaterPage
tests pass unchanged.
